feature: root level server time query
This PR adds the 'time' field to the root level query, allowing a consistent server time to be returned in any query, which makes datetime calculations easier as you don't have to rely o the client device time.

This PR also attempts to bring the Root model more in line with the other models, with the same folder structure.